requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
langgraph>=0.0.15
langchain>=0.1.0
langgraph-checkpoint-postgres>=0.1.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0